Moved generate_viz  out of main, and into viz
